Rebase from 'release/indigo/robot_calibration'

+++ src/optimization_params.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2014 Fetch Robotics Inc.
+ * Copyright (C) 2014-2015 Fetch Robotics Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -21,7 +21,7 @@
 namespace robot_calibration
 {
 
-OptimizationParams::OptimizationParams() : 
+OptimizationParams::OptimizationParams() :
   base_link("base_link")
 {
 }
@@ -38,7 +38,7 @@ bool OptimizationParams::LoadFromROS(ros::NodeHandle& nh)
     nh.getParam("free_params", names);
     ROS_ASSERT(names.getType() == XmlRpc::XmlRpcValue::TypeArray);
 
-    for (size_t i = 0; i < names.size(); ++i)
+    for (int i = 0; i < names.size(); ++i)
     {
       free_params.push_back(static_cast<std::string>(names[i]));
     }
@@ -52,7 +52,7 @@ bool OptimizationParams::LoadFromROS(ros::NodeHandle& nh)
     nh.getParam("free_frames", free_frame_params);
     ROS_ASSERT(free_frame_params.getType() == XmlRpc::XmlRpcValue::TypeArray);
 
-    for (size_t i = 0; i < free_frame_params.size(); ++i)
+    for (int i = 0; i < free_frame_params.size(); ++i)
     {
       FreeFrameParams params;
       params.name = static_cast<std::string>(free_frame_params[i]["name"]);
@@ -74,7 +74,7 @@ bool OptimizationParams::LoadFromROS(ros::NodeHandle& nh)
     nh.getParam("models", model_params);
     ROS_ASSERT(model_params.getType() == XmlRpc::XmlRpcValue::TypeArray);
 
-    for (size_t i = 0; i < model_params.size(); ++i)
+    for (int i = 0; i < model_params.size(); ++i)
     {
       Params params;
       params.name = static_cast<std::string>(model_params[i]["name"]);
@@ -92,7 +92,7 @@ bool OptimizationParams::LoadFromROS(ros::NodeHandle& nh)
     nh.getParam("error_blocks", error_params);
     ROS_ASSERT(error_params.getType() == XmlRpc::XmlRpcValue::TypeArray);
 
-    for (size_t i = 0; i < error_params.size(); ++i)
+    for (int i = 0; i < error_params.size(); ++i)
     {
       Params params;
       params.name = static_cast<std::string>(error_params[i]["name"]);
